Uses the since method to write custom messages to the user...
It adds the since method to the first `it` block of the first file of part 1.

The first `it` block contains three `expect` methods, where the first and the third `expect` methods do not fail, therefore they don't throw an error.

The second `expect` block will fail and throw the custom message.

The custom message can be verified in the console and in the reporter.

The reporter can be seen inside the `karma-result.json` file.